refactor(gpt-oss): move NKI kernels into kernels/nki/
Extract the hand-written NKI RMSNorm (nki_rmsnorm_kernel + stream_shuffle_
broadcast helper) out of kernels/sampling.py into a dedicated kernels/nki/
package, so the low-level NKI code is separated from the numpy-traced kernels.
sampling.py now imports it from .nki.rmsnorm. Behavior unchanged (default
USE_NKI_RMSNORM=True); verified generation still produces "Paris" at ~53 tok/s.
Documented the new subfolder in the README Files table.
